requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.8
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
